refactor: widgets를 FSD segments로 분리
각 widget을 lib, model, ui 세그먼트로 세분화하여 관심사 분리
