+++ src/widgets/post-form-dialog/index.ts
@@ -1,0 +1,1 @@
+export { PostFormDialogWidget } from './ui/PostFormDialogWidget'
