+++ src/widgets/posts-search/index.ts
@@ -1,0 +1,1 @@
+export { PostsSearchWidget } from './ui/PostsSearchWidget'
