+++ src/widgets/posts-table/index.ts
@@ -1,0 +1,1 @@
+export { PostsTableWidget } from './ui/PostsTableWidget'
