+++ src/widgets/user-profile-modal/index.ts
@@ -1,0 +1,1 @@
+export { UserProfileModalWidget } from './ui/UserProfileModalWidget'
