+++ src/widgets/user-profile-modal/ui/UserProfileModalWidget.tsx
@@ -1,19 +1,8 @@
-import { useUserQuery } from '@features/user/get-single-user'
-import { usePostsUI } from '@shared/store/posts-ui'
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from '@shared/ui/Dialog'
+import { useUserProfileData } from '../model/useUserProfileData'
 
 export const UserProfileModalWidget = () => {
-  const { selectedUserId, showUserModal, setShowUserModal, setSelectedUserId } = usePostsUI()
-
-  // ID로 사용자 데이터 조회
-  const { data: selectedUser, isLoading } = useUserQuery(selectedUserId!, {
-    enabled: !!selectedUserId && showUserModal,
-  })
-
-  const handleCloseModal = () => {
-    setShowUserModal(false)
-    setSelectedUserId(null)
-  }
+  const { selectedUser, isLoading, showUserModal, handleCloseModal } = useUserProfileData()
 
   return (
     <Dialog open={showUserModal} onOpenChange={handleCloseModal}>
